feat: refactor CIA content service and hooks to enhance type safety and maintainability; introduce default security icons and streamline ROI metrics handling
